uploader/requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
watchdog>=4.0.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
schedule>=1.2.2
pytest>=8.0.0
pytest-mock>=3.14.0
